+++ docs/changelog.md
@@ -85,7 +85,7 @@
         * Config option to control chance of extra anvil damage when hammering a bloom
             * Config value: `module.tech.Basic.cfg` -> `GRANITE_ANVIL` -> `BLOOM_EXTRA_DAMAGE_CHANCE`
             * Config value: `module.tech.Basic.cfg` -> `IRONCLAD_ANVIL` -> `BLOOM_EXTRA_DAMAGE_CHANCE`
-
+    * When an anvil breaks with a bloom on it, the bloom will be placed in the world where the anvil was [(#129)](https://github.com/codetaylor/pyrotech/issues/129)
 ---
 
 1.3.12